[10-refactoring] model -> spec 이름 변경
