+++ .idea/misc.xml
@@ -3,8 +3,7 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Users/user/AndroidStudioProjects/pozi/app/src/main/res/layout/activity_main.xml" value="0.17980072463768115" />
-        <entry key="..\:/Users/user/AndroidStudioProjects/pozi/app/src/main/res/layout/item_photobooth_detail_viewpager.xml" value="0.4" />
+        <entry key="..\:/Users/user/AndroidStudioProjects/pozi/app/src/main/res/layout/item_photobooth_detail_viewpager.xml" value="0.17980072463768115" />
       </map>
     </option>
   </component>
