+++ app/src/main/java/com/example/pozi_android/data/remote/network/LocationRes.kt
@@ -1,6 +1,6 @@
 package com.example.pozi_android.data.remote.network
 
-import com.example.pozi_android.data.remote.model.Locations
+import com.example.pozi_android.data.remote.spec.Locations
 import com.google.gson.annotations.SerializedName
 
 class LocationRes {
