+++ app/src/main/java/com/example/pozi_android/data/remote/spec/Locations.kt
@@ -1,22 +1,21 @@
-package com.example.pozi_android.data.remote.model
+package com.example.pozi_android.data.remote.spec
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
 
 data class Locations(
-    @SerializedName("id")
-    @Expose
+    @SerializedName("id") //이거는 받아오는 순서대로 넣어주는 라이브러리 사용하기
     val id: Int,
+
     @SerializedName("address")
-    @Expose
     val address: String,
+
     @SerializedName("lat")
-    @Expose
     val lat: Double,
+
     @SerializedName("lng")
-    @Expose
     val lng: Double,
+
     @SerializedName("name")
-    @Expose
     val name: String
 )
