+++ app/src/main/java/com/example/pozi_android/domain/mapper/empty.kt
@@ -1,0 +1,4 @@
+package com.example.pozi_android.domain.mapper
+
+object empty {
+}
